Client: Move socket io connection initiation and a few other related effects from messages-page component to main app component
